# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "KAVACH API",
        "description": "Risk, Fraud and Regulatory Intelligence Copilot for Indian Banking",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/rings": {
            "get": {
                "tags": [
                    "Rings"
                ],
                "summary": "List Rings",
                "description": "List mule rings",
                "operationId": "list_rings_api_rings_get",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "default": 1,
                            "title": "Page"
                        }
                    },
                    {
                        "name": "page_size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1,
                            "default": 20,
                            "title": "Page Size"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RingListResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/rings/{ring_id}": {
            "get": {
                "tags": [
                    "Rings"
                ],
                "summary": "Get Ring Detail",
                "description": "Get detailed information for a specific ring",
                "operationId": "get_ring_detail_api_rings__ring_id__get",
                "parameters": [
                    {
                        "name": "ring_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Ring Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RingDetailResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "RingDetailResponse": {
                "properties": {
                    "ring": {
                        "$ref": "#/components/schemas/RingResponse"
                    },
                    "members": {
                        "items": {
                            "type": "object"
                        },
                        "type": "array",
                        "title": "Members"
                    },
                    "transactions": {
                        "items": {
                            "type": "object"
                        },
                        "type": "array",
                        "title": "Transactions"
                    }
                },
                "type": "object",
                "required": [
                    "ring",
                    "members",
                    "transactions"
                ],
                "title": "RingDetailResponse",
                "description": "Ring detail response model"
            },
            "RingListResponse": {
                "properties": {
                    "rings": {
                        "items": {
                            "$ref": "#/components/schemas/RingResponse"
                        },
                        "type": "array",
                        "title": "Rings"
                    },
                    "total": {
                        "type": "integer",
                        "title": "Total"
                    },
                    "page": {
                        "type": "integer",
                        "title": "Page"
                    },
                    "page_size": {
                        "type": "integer",
                        "title": "Page Size"
                    },
                    "total_pages": {
                        "type": "integer",
                        "title": "Total Pages"
                    }
                },
                "type": "object",
                "required": [
                    "rings",
                    "total",
                    "page",
                    "page_size",
                    "total_pages"
                ],
                "title": "RingListResponse",
                "description": "Ring list response model"
            },
            "RingResponse": {
                "properties": {
                    "ring_id": {
                        "type": "string",
                        "title": "Ring Id"
                    },
                    "ring_name": {
                        "type": "string",
                        "title": "Ring Name"
                    },
                    "member_count": {
                        "type": "integer",
                        "title": "Member Count"
                    },
                    "total_volume_inr": {
                        "type": "number",
                        "title": "Total Volume Inr"
                    },
                    "risk_score": {
                        "type": "number",
                        "title": "Risk Score"
                    },
                    "status": {
                        "type": "string",
                        "title": "Status"
                    }
                },
                "type": "object",
                "required": [
                    "ring_id",
                    "ring_name",
                    "member_count",
                    "total_volume_inr",
                    "risk_score",
                    "status"
                ],
                "title": "RingResponse",
                "description": "Ring response model"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
        }
    }
}